feat(plugin) add support proxy server for all plugins
