+++ package.json
@@ -26,6 +26,9 @@
   },
   "dependencies": {
     "boxen": "^8.0.1",
+    "chalk": "^4",
+    "https-proxy-agent": "^7.0.6",
+    "node-fetch": "^3.3.2",
     "update-notifier": "^7.3.1"
   }
 }
