+++ packages/generator-amazon-apigateway/src/test/plugin.test.ts
@@ -7,7 +7,7 @@ import { vi } from 'vitest';
 import { GetExportCommand, GetRestApisCommand } from '@aws-sdk/client-api-gateway';
 
 // Add mock for the local checkLicense module
-vi.mock('../checkLicense', () => ({
+vi.mock('../../../../shared/checkLicense', () => ({
   default: () => Promise.resolve(),
 }));
