+++ packages/generator-asyncapi/package.json
@@ -43,7 +43,6 @@
     "fs-extra": "^11.2.0",
     "glob": "^11.0.0",
     "gray-matter": "^4.0.3",
-    "https-proxy-agent": "^7.0.6",
     "js-yaml": "^4.1.0",
     "lodash": "^4.17.21",
     "minimist": "^1.2.8",
