+++ packages/generator-openapi/src/test/plugin.test.ts
@@ -7,7 +7,7 @@ import { vi } from 'vitest';
 import { existsSync } from 'node:fs';
 
 // Add mock for the local checkLicense module
-vi.mock('../utils/checkLicense', () => ({
+vi.mock('../../../../shared/checkLicense', () => ({
   default: () => Promise.resolve(),
 }));
